requirements.txt
customtkinter
tk
pillow
ultralytics
torch 
torchvision
tkvideoplayer